feat: implement sales functionality
two new tables are created: Sales and SaleProducts. service and
controller for *partial* sales functionality is created

+++ src/main/java/com/stasy/api/dtos/ProductDTO.java
@@ -2,5 +2,7 @@ package com.stasy.api.dtos;
 
 import com.stasy.api.domain.product.ProductCategory;
 
-public record ProductDTO(String name, String manufacturer, ProductCategory category, double price, long quantity) {
+import java.math.BigDecimal;
+
+public record ProductDTO(String name, String manufacturer, ProductCategory category, BigDecimal price, long quantity) {
 }
